Function thirtyseconds::strings::is_anagram

pub fn is_anagram(sentence1: String, sentence2: String) -> bool

Checks if a string is an anagram of another string (case-insensitive, ignores spaces, punctuation and special characters).

Remove any whitespace using replace().

Lowercase all the string using to_lowercase().

Compare the lengths of the both, return False if they are not equal.

If they have the same length, sort the string and compare.

Examples

Basic usage:

assert!(is_anagram("iceman".to_string(), "cinema".to_string()));